==> 0_check_T60_companies.ipynb <==
import requests, json
from requests.auth import HTTPBasicAuth
# import credentials
import json
import pandas as pd
import numpy as np
import os  # Ensure this is imported at the top

country_list = ["JAPAN", "EURO", "UK", "USA", "AUS", "INDIA"] # "INDIA", "JAPAN", "EURO", "UK", "USA", "AUS"
# Define the base directory where you want to save the files
# base_directory = r"C:\Users\60848\OneDrive - Bain\Desktop\Project_Genome\global_platform_data"
BASE_DIR = os.path.join(os.path.abspath(os.path.curdir), "..")
GLOBAL_MARKET_DIR = os.path.join(BASE_DIR, "Global_market_constituents")
GLOBAL_DATA_DIR = os.path.join(BASE_DIR, "global_platform_data")

all_companies = []
all_tickers = []
all_full_tickers = []
all_data = []
for c in range(len(country_list)):

    country_import = country_list[c] # AUS / USA / INDIA / JAPAN / euro / UK

    # Import mapping data
    if country_import == "USA":
        # mapping_data = pd.read_csv(r"C:\Users\60848\OneDrive - Bain\Desktop\Project_Genome\Global_market_constituents\Company_list_GPT_SP500.csv", encoding='cp1252')
        mapping_data = pd.read_csv(os.path.join(GLOBAL_MARKET_DIR, "Company_list_GPT_SP500.csv"), encoding='cp1252')
    if country_import == "AUS":
        # mapping_data = pd.read_csv(r"C:\Users\60848\OneDrive - Bain\Desktop\Project_Genome\Global_market_constituents\Company_list_GPT_AUS.csv", encoding='cp1252')
        mapping_data = pd.read_csv(os.path.join(GLOBAL_MARKET_DIR, "Company_list_GPT_AUS.csv"), encoding='cp1252')
    if country_import == "INDIA":
        # mapping_data = pd.read_csv(r"C:\Users\60848\OneDrive - Bain\Desktop\Project_Genome\Global_market_constituents\Company_list_GPT_Nifty.csv", encoding='cp1252')
        mapping_data = pd.read_csv(os.path.join(GLOBAL_MARKET_DIR, "Company_list_GPT_Nifty.csv"), encoding='cp1252')
    if country_import == "JAPAN":
        # mapping_data = pd.read_csv(r"C:\Users\60848\OneDrive - Bain\Desktop\Project_Genome\Global_market_constituents\Company_list_GPT_Nikkei.csv", encoding='cp1252')
        mapping_data = pd.read_csv(os.path.join(GLOBAL_MARKET_DIR, "Company_list_GPT_Nikkei.csv"), encoding='cp1252')
    if country_import == "EURO":
        # mapping_data = pd.read_csv(r"C:\Users\60848\OneDrive - Bain\Desktop\Project_Genome\Global_market_constituents\Company_list_GPT_Euro.csv", encoding='cp1252')
        mapping_data = pd.read_csv(os.path.join(GLOBAL_MARKET_DIR, "Company_list_GPT_Euro.csv"), encoding='cp1252')
    if country_import == "UK":
        # mapping_data = pd.read_csv(r"C:\Users\60848\OneDrive - Bain\Desktop\Project_Genome\Global_market_constituents\Company_list_GPT_FTSE.csv", encoding='cp1252')
        mapping_data = pd.read_csv(os.path.join(GLOBAL_MARKET_DIR, "Company_list_GPT_FTSE.csv"), encoding='cp1252')

    companies = mapping_data["Company name"].unique().tolist()
    tickers = mapping_data["Ticker"].unique().tolist()
    full_tickers = mapping_data["Full Ticker"].unique().tolist()
    mapping_data = mapping_data[["Ticker", "Full Ticker", "Company name", "Sector_new"]]
    mapping_data = mapping_data.drop_duplicates()
    mapping_data["Country"] = country_import
    all_data.append(mapping_data)
    
    all_companies.extend(companies)
    all_tickers.extend(tickers)
    all_full_tickers.extend(full_tickers)

df_all = pd.concat(all_data)

len(all_companies), len(all_tickers), len(all_full_tickers)

remove_duplicates_companies = []
remove_duplicates_tickers = []
remove_duplicates_full_tickers = []

# remove duplicates companies from dataset
mapping_data = pd.read_csv(os.path.join(GLOBAL_MARKET_DIR, "Company_list_GPT_TS60_.csv"), encoding='cp1252')
print(mapping_data.shape)
companies = mapping_data["Company name"].unique().tolist()
tickers = mapping_data["Ticker"].unique().tolist()
full_tickers = mapping_data["Full Ticker"].unique().tolist()

for c in companies:
    if c in all_companies:
        remove_duplicates_companies.append(c)
for t in tickers:
    if t in all_tickers:
        remove_duplicates_tickers.append(t)
for ft in full_tickers:
    if ft in all_full_tickers:
        remove_duplicates_full_tickers.append(ft)

final_data = mapping_data[~mapping_data["Company name"].isin(remove_duplicates_companies)]
print(final_data.shape)
final_data = final_data[~final_data["Ticker"].isin(remove_duplicates_tickers)]
print(final_data.shape)
final_data = final_data[~final_data["Full Ticker"].isin(remove_duplicates_full_tickers)]
print(final_data.shape)

df_all[df_all["Full Ticker"].isin(remove_duplicates_full_tickers)]

print(mapping_data.shape)
mapping_data

final_data = final_data.reset_index(drop=True)
final_data["Ticker"] = final_data["Full Ticker"].apply(lambda x: x.split(":")[1])
final_data.to_csv(os.path.join(BASE_DIR, "Company_list_GPT_TS60_clean.csv"), index=False)
final_data

remove_duplicates_tickers

remove_duplicates_full_tickers

mapping_data = pd.read_csv(os.path.join(GLOBAL_MARKET_DIR, "Company_list_GPT_SP500.csv"), encoding='cp1252')
mapping_data.shape


mapping_data["Company name"].unique().tolist()

mapping_data

mapping_data = pd.read_csv(os.path.join(GLOBAL_MARKET_DIR, "Company_list_GPT_TS60_.csv"), encoding='cp1252')
mapping_data.shape

mapping_data
